graph: Add a new API for the execution graph
This patch adds interfaces for a new API for the graph. The ITmfGraph
interface is the main entry point for the whole API, edges and vertex
will be created through this interface, instead of directly, so the
exact class implementations are not visible to the API consumers.

This will allow various implementations, either in memory with the
current graph or eventually on disk.

Change-Id: Ibb68deadc4657fc7e840064c51824df7d0fcbadd
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/182711
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>