graph: Deprecate legacy API and update code to use the new ITmfGraph
Deprecate the TmfGraph, TmfVertex and TmfEdge classes and update the
code of the graph builder module and critical path to use the new
ITmfGraph API. For now, the legacy are simply wrapped in an ITmfGraph
implementation, but various implementations can be added to save the
graph on disk (for later).

Unit tests are updated to use the new API as well.

Deprecation errors are turned into warnings in the graph.core.* projects
to avoid errors because the in-memory graph still use the legacy classes
that should be moved to internal once the deprecated classes are
removed.

[Deprecated] The TmfGraph, TmfVertex and TmfEdge are deprecated in favor
of the ITmfGraph interface.
[Changed] The analysis.graph and LTTng execution graph classes now use
the ITmfGraph API.

Change-Id: I744dbc21a33f901ab2386c6501db06b03c94d3aa
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/182723
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>