graph: Decouple OS specific edges replacing EdgeType with contextState
The Edge type can designate multiple attributes related to the edge: the
edge orientation, direction, underlying state and the state it will be
evaluated as when running the critical path algorithm. This can be
confusing so when refactoring this module, I renamed it to contextState.

The ContextState is not coupled to the linux OS, and allows the mapping
between the context in which it is used (operating systems, clusters,
etc...) and the state in the critical path algorithm (passing or
blocking).

The state in the critical path algorithm is defined by the TmfEdgeState.
The contextState is defined by the ITmfEdgeContextState.

Change-Id: I406e2544a0245e30040e5c284818a283419a4162
Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/192194
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>