timing.core: Search in segment store table data provider
In the current segment store table data provider when searching for
segments by applying some search filter, the fetched segments that
matches the filter criteria were not highlighted, this change highlights
the segments that are matching the search filter so finding the desired
segment will be easier. Also, it supports seeking to the next or
previous matching segment when a search direction is provided.

Change-Id: I8e14e7204833143267998d8245e2a395b98959db
Signed-off-by: Kyrollos Bekhet <kyrollos.bekhet@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/194451
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Hoang Thuan Pham <hoangpham.eclipse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>