timing.core: Support for experiment in segment store table data provider
Before this commit the segment store table data provider only displayed
the segment for single traces and did not support experiments. With this
change the segments for all traces are returned by the data provider.

Fix an error in the JSON file generated by Trace Compass traces because
the analysis name of SegmentStoreAnalysisModule was not defined. The
error was fixed by assigning the name of the SegmentStoreAnalysisModule
to the name of the analysis in progress.

Change-Id: I00276106235a4c74b1d75daef29034ebcfb90601
Signed-off-by: Kyrollos Bekhet <kyrollos.bekhet@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/194955
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>