xml: Disable sonar given FEATURE_SECURE_PROCESSING
Add the use of XMLConstants.FEATURE_SECURE_PROCESSING where previously
missing for the XML parsers. Disable the related Sonar flags [1] which
cannot recognize this as the mitigating (currently known) fix for them.

In [2], FEATURE_SECURE_PROCESSING cannot be set, as that leads to schema
construction failures. Add a comment explaining this further within [2].

Eclipse (by default) fails to recognize the corresponding, more specific
@SuppressWarnings annotated token unfortunately. Hence use the //NOSONAR
approach despite its act of cancelling all potential Sonar flags for the
whole line on which it sits.

Foster the benefit of removing [1] this way from both the centralized
and local Sonar installations (e.g. SonarCloud, and SonarLint in IDEs),
for everyone in the project. Disabling such a security warning is done
here only because of addressing it through FEATURE_SECURE_PROCESSING,
or at least having tried to ([2]); the flag shall be kept otherwise.

[1] "Disable access to external entities in XML parsing." (Sonar's.)
[2] o.e.t.internal.tmf.analysis.xml.core.module.XmlUtils.xmlValidate

Change-Id: Iff26934d01964d525dbae773f4a6c94817f9970f
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/195681
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>