ui: Set //NOSONAR for regexes in these few classes
Disable this local SonarLint warning about its java:S5998 rule [1,2]
below, which shows as the other rule [3] in SonarCloud. Do so as this is
basically not an issue for Trace Compass today.

Eclipse fails to recognize the corresponding, more specific
@SuppressWarnings token. Hence use //NOSONAR despite its act of
cancelling all potential Sonar flags for the whole line. Foster the
benefit of removing [2,3] this way from both SonarCloud and SonarLint
(the latter in local IDEs).

LttngRelaydConsumer, however, didn't show [1,2] locally but does have
[3] that showed in SonarCloud. Same for LTTngControlServiceConstants,
X11ColorUtils, IMappingFile and TmfAbstractToolTipHandler (in order of
SonarCloud appearance).

[1] Refactor this repetition that can lead to a stack overflow for large inputs.
[2] https://sonarcloud.io/organizations/eclipse/rules?open=java%3AS5998&rule_key=java%3AS5998
[3] https://sonarcloud.io/organizations/eclipse/rules?open=java%3AS5852&rule_key=java%3AS5852

Change-Id: Ibc1f4aa43a7fc48d5d5437c10483323a0bd217aa
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/195750
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>