segmentstore: Add "Previous" test alongside Next's
Add the testDataProviderFetchLineWithSearchPrevious method to
SegmentStoreTableDataProviderTest, based on the existing
testDataProviderFetchLineWithSearch one for the Next case.

Apply default Eclipse formatting to the applicable line while in there.

This test might not behave as expected, so may require amends, say in
the context of [1]. It is passing currently but likely out of missing
checks or more complete test data.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=580674

Change-Id: I103d334b26c09f1c811b302243fec17d7df0d46c
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/195979
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>