counters: Fix AbstractCounterAspect#resolve's API
Commit ac4cf8f2db replaced AbstractCounterAspect#resolve returning a
Long with now returning a Number. The latter is either a Double or a
Long depending on the counter type. Slightly amend the related code to
make this properly explicit, enum/switch handling wise.

This change also clarifies the related @return javadoc accordingly,
while clarifying the next release news-and-noteworthy with the footer
down below. Mention the still existing possibility of returning null,
in that javadoc solely.

Preserve the ability to release in a minor fashion next time, by means
of replacing commit ac4cf8f2db's 3.0.0 (an otherwise major bump) with
2.1.0 in the related files.

Add a commented Eclipse API filter to remove the two build errors caused
by not stepping the major version this way. Deem this API change as a
non-breaking one, again minor only, manually thus on purpose; per above.

[Changed] AbstractCounterAspect#resolve now returns Number (Double|Long)

Change-Id: I2e0859f9232399d9be9383ac6d9e95fb8e702ec3
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/196395
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>