Bug 581074: Fix support of Javascript state values for Java>=15 for XML
Nashorn scripting engine was removed starting with OpenJDK 15. Therefore
Javascript state values in XML analysis didn't work anymore when
using Java 15 or newer.

This patch adds support of Mozilla's Rhino Javascript engine so that
this XML Javascript feature can be used with Java 15 or newer.

Default to engine "rhino" if "nashorn" is not available.

[Fixed] support of Javascript state values for Java>=15 in XML analysis
[Added] support of Mozilla's Rhino Javascript engine for XML analysis

Change-Id: I8806d7f31284d1704d59ebbb120649b91b9cfdc1
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/197112
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>