analysis: Initialize a new 'callstack' plugin core
Base the new minimal set of plugin files on sibling profiling.core ones.

This new plugin is to incrementally hold the callstack implementation to
bring in from Incubator. Goal being, to main-line the latter's callstack
in Trace Compass, as an improved implementation over the current, more
limited one.

Scaffold the incubated callstack alongside the 'profiling' one this way,
before potentially deprecating the latter once proven as a subset.

When run in Eclipse, ActivatorTest's getDefault and getAbsoluteFilePath
methods are not covered locally. Keep them still to match with
profiling.core's own lack of coverage for those, consistently, for this
change scope.

Change-Id: Iaad2fd837c3d58f4b6c6a0514dfdd7cfd0d5cee7
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/197869
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>