Remove MAINTAINERS.md file from repository root directory
The idea of MAINTAINERS.md file was to indicate to contributors which
component is maintained by which committer and to know who to ask for
a review. However, in practice this didn't play out as expected. Instead
each committer is able to review any of the contribution.

If a committer is less familiar with a component, that committer can
request a review by another committer.

Moreover, the MAINTAINERS.md is outdated and the referenced list of
committers is not up-to-date.

Contributors can find the list of active committers here:
https://projects.eclipse.org/projects/tools.tracecompass/who

Change-Id: I5a117fca260d8e5b91c6978dd716087a6c691926
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/199291
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>