callstack: Move these to callstack.core.instrumented
As a way to keep reducing the size of callstack.core's WIP flat list.

Base this latest packaging on the current relationships among these
objects, as started with the base commits.

Move CallStackGroupDescriptor and EdgeStateValue from core.base to
core.instrumented, now that they are solely used by the latter.

Change-Id: Ie5fbcc438a122caa4a5faf322873db50e7c224b5
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/199466
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>