tmf.ui: Cache size and bounds on resize or move
It is impossible to resize while in a paint operation, so the getSize()
and getBounds() operation can be cached. In doing this, each link does
not call the getSize() and getBounds() functions twice per draw.

Accelerates from 15s/draw to 500ms/draw on "bar code" style images.

This is very evident on trace compass traces, especially while using
LogLevel#Finest

[Changed] Improve performance of timegraph with many links/markers

Change-Id: I7e228056ef28a988d4c8d7fb3803f16e1b1bfb6d
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/201710
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>