gdbtrace.core: handle selection with heterogenous traces
Check that a child trace is a GdbTrace before casting.

Change comparison from object.equals(other) to Objects.equals(a,b)
This avoids exceptions. This should not change the functioning, but
reduce the false positive errors (NPE) in the console. The exception
before was aborting the if statement on false.

Change-Id: I1636e9cb856e5fbf69a33d9e6a8e6cbaa539219e
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/201860
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>