linux.core: Generalize ResourcesStatusDataProvider
The original ResourcesStatusDataProvider was parameterized to require
a KernelAnalysisModule. Now ResourcesStatusDataProvider is requires a
TmfStateSystemAnalysisModule, making it available for other analysis
modules such as DataDrivenAnalysisModule. The expected information
can then be provided by other analysis module with state systems
in the same structure.

Note: this class is internal and thus would need a factory to take
advantage of the increased modularity.

[Changed] ResourceStatusDataProvider depends on TmfStateSystemAnalysisModule

Signed-off-by: Siwei Zhang <manbearpig10086@gmail.com>
Change-Id: I45030de7319bfab4092e49cf10a92309b3bf0724
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/201928
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>