Bug 581642: Correctly handle critical path file
Previously the history tree file was never closed and it also never
was reopened correctly. In Windows, this resulted in errors because
the process was not able to delete the file while it was still
opened. This patch adds the capability of comparing the worker to
know if we need to delete the previous critical path or not.

There was also an error where when a new thread was followed, the
previous critical path was set to null without closing the files.
This patch corrects that issue by always disposing the critical
path before setting it to null.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Change-Id: Ia8a504835c5c51c3d25243afbee0c63b656d0690
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/202590
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>