tmf.ui: Correct duplicate entries error
When the provider returns multiple time the same entries, they will
be already added to the parent and so they should not be counted
twice. If they are counted multiple times, when another child is
added, it will be added to the wrong index and will result in an
exception.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Change-Id: Ib34ec74348690bb264437b89a86a13b35a0a3d7f
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/202594
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>