ctf.core: Proper initialization of a CTF2 Trace
Working on the reading of a packet header and a data file to load the
trace without error. Will automatically be tested by JUnit tests and
can manually be tested using [1]. The expected result is a trace with
only one event type ('string') and every event should have the same
timestamp. More complex traces cannot be tested until clock
functionality is added.

[1] https://review.lttng.org/plugins/gitiles/babeltrace/+/refs/heads/ctf2/tests/data/ctf-traces/2/succeed/succeed1/

Change-Id: I2f644e3d325cbe6846442cebb775a8faba60a2ec
Signed-off-by: Sehr Moosabhoy <sehr.moosabhoy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203295
Tested-by: Marco Miller <marco.miller@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>