timing.ui: no tooltip when no data is hovered
The binary search here will return a negative number if and only if
there is no value found. If there is no value found, that means the
mouse is hovering a section of the graph where there is no data. In that
case, no tooltip should be displayed.

To test that, the user should use the Density Viewer of any analysis
and hover on the leftmost side of the graph. It might be a quite small
area, but zooming on the leftmost side of the view can make that error
appear often.

The previous behaviour ends up throwing an error, because if the binary
search returns -1, the calculated index is -1 which is out of bounds.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Change-Id: Ic7f312babc31367af3182ff14f235e17e5205b51
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203431
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>