linux: Move SWSLatency analysis to Trace Compass core
The SWSLatency computes the latency between the sched_wakeup event and
the sched_switch event for each thread. It is a subclass of
AbstractSegmentStoreAnalysisEventBasedModule.

This commit moves the SWSLatency analysis from the incubator to Trace
Compass core by copying the necessary files from
org.eclipse.tracecompass.incubator.internal.kernel.core.swslatency to
org.eclipse.tracecompass.analysis.os.linux.core.swslatency and making
necessary changes to existing configuration files. It also moves unit
tests for the analysis to Trace Compass core.

To test the analysis:

[1] Run the unit tests in
org.eclipse.tracecompass.analysis.os.linux.core.tests.swslatency
as J-unit plug-in tests. Confirm that they all pass. The goal is to test
that the analysis runs and provides expected output.
[2] Run the unit tests in
org.eclipse.tracecompass.integration.core.tests.dataproviders
as J-unit plug-in tests. Confirm that they all pass. Adding the new
analysis generates new DataProviderDescriptor(s). This is to make sure
that the existing test in this package is updated.
[3] Open Trace Compass. Open a kernel trace. Make sure that the
Scheduler Wakeup to Scheduler Switch Latency analysis is visible and not
crossed out under Views in the Project Explorer. It is normal to have no
output under the analysis because they are not yet included in this
commit. This is required to make sure that the analysis is exposed to
users in the UI of the Trace Compass application.

To be done in subsequent patches:

[1] Migrate outputs of the SWSLatency analysis
[2] Renaming the SWSLatency analysis to
SchedulerWakeupToSchedulerSwitchLatency to make the name of the analysis
easier to understand
[3] Add documentation
[4] Deprecate the SWSLatency analysis in the incubator

[Added] SWSLatency (SchedulerWakeupToSchedulerSwitchLatency) analysis

Change-Id: I7e105c4bae2042fd230d43eee9e97972113b8529
Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203530
Tested-by: Marco Miller <marco.miller@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>