releng: Bump mvn build plugin versions in pom.xml
Bump the single maven build plugin versions in ./pom.xml that are left
to upgrade, except the ones mentioned below.

Do not bump the maven-install-plugin version now, as doing so fails the
build currently. -Same for the tycho version which requires another
focused task, being even less trivial to address.

Change-Id: I331055391f84eeebcebd74fd4b12f9e7af4860c8
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203617
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>