ctf.core: Implement parsing of clock fragments
Add logic to parse CTF2 clock fragments and therefore allow for
timestamp readings in CTF2. This implementation can be tested with a
modified version of [1] that will be shared (updated to include
CTF2-SPECRC-7.0 terminology) and can be compared with the equivalent
CTF1 trace [2].
Babeltrace has yet to share tests that include clock origins, so those
cannot be tested yet. An additional change was made to remove the
datastream TODO that was left over from a previous patch.

[1]https://review.lttng.org/plugins/gitiles/babeltrace/+/refs/heads/ctf2/tests/data/ctf-traces/2/succeed/barectf-event-before-packet/

[2]https://review.lttng.org/plugins/gitiles/babeltrace/+/refs/heads/ctf2/tests/data/ctf-traces/1/succeed/barectf-event-before-packet/

Change-Id: I46c2de5305b72ce8e4ef4834708b5887bd811bfc
Signed-off-by: Sehr Moosabhoy <sehr.moosabhoy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203658
Tested-by: Marco Miller <marco.miller@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>