ctf.core: Update integer and blob declarations to have a role field
The CTF2 role field was added in the declaration class and new
constructors were made for the type declarations that can have roles:
integers and blobs. Tests for blobs and methods using roles were
updated. A new method was also added to determine whether a trace is
CTF2. Since the IntegerDeclaration constructor was deprecated, all of
the tests & files that use it were also updated.

[added] roles for CTF events

Change-Id: I54bce6150539bbb2ff380cb12c4b39c7045e661e
Signed-off-by: Sehr Moosabhoy <sehr.moosabhoy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203774
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>