ctf.core: Add support for CTF2 enumerations
Add CTF2 parsing of fixed length signed enumerations. This patch does
not yet completely cover variable length enumerations. Testing of this
patch is delayed for now until the following variant patch is approved.
JUnit tests for this patch should be added with CTF2 updated tests in
the variantDefinition, variantDeclaration, and EventHeaderDeclaration
files.

Change-Id: I422c2669931c67f5b293abafb16f10a3635d158d
Signed-off-by: Sehr Moosabhoy <sehr.moosabhoy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203783
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>