ctf.core: Add CTF2 support for variants
Added support for CTF2 variant field classes. This also slightly changed
the structure field metadata nodes. A change was also made to the user
attributes field in CTFJsonMetadataNode. This patch can be tested using
a CTF2-7.0 compliant version of [1] and these traces will also be added
to the test-traces repo soon.

[1]https://review.lttng.org/plugins/gitiles/babeltrace/+/refs/heads/ctf2/tests/data/ctf-traces/2/succeed/trace-with-index/

Change-Id: I603abc533f0281e071b4bb9ff0ca5538d4da1283
Signed-off-by: Sehr Moosabhoy <sehr.moosabhoy@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/203784
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>