Fix TraceCompass.setup problems
Do not install org.eclipse.pde.api.tools.ee.feature.feature.group
because it is no longer available in the Platform and it makes the p2
director task fail.

Redirect the setup location using https because that's what's used in
Oomph's project catalog so if that's not specified, one can't edit the
local version in the clone and test it locally.

Change-Id: I4829e82e8cc58437338e2c46757a2c763a99a973
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/204053
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/204106
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>