tmf: Bug 582385: Fix StackOverflowException in EventDensity view
The this patch fixes a StackOverflowException in the
HistogramDataProviderFactory when a TmfTrace extension filters out the
framework, default TmfStatisticsModule. The factory now returns null
instead of falsely calling the TmfTreeXYCompositeDataProvider which
ends-up in an endless recursion.

Even though it's a default analysis module, extensions can choose to
remove it and the framework view should not cause such an exception.

[Fixed] Bug 582385: Fix StackOverflowException in EventDensity view

Change-Id: Id904867f7559bcd6cc93588fffabe47c7a5a1c77
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/204200
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>