ctf: Support overlapping enum values
Support overlapping of both single values and values ranges in the enum
declaration. The internal structure is changed from a TreeSet to a
Multimap. Overlapping can occur for the same key in the multimap
(overlapping single values) or for different keys in the multimap
(overlapping value ranges).

The query result label is the same when no overlap occurs, but when
there is overlap, the matching labels are returned as [l1, l2, ...].
This format can also be returned in the lookup table.

Add unit tests for overlapping and adding without specifying range.

[Added] ctf: Support overlapping enum values

Change-Id: I19fe159727b41fea324041ce0b20edcfd2cae36e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/204295
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>